#956 Phase 1: extract cos/ecn.rs from tx.rs
Establishes the userspace-dp/src/afxdp/cos/ submodule. Phase 1
extracts the most self-contained CoS subsystem (ECN marking) and
sets up the pattern that subsequent phases (admission, token
bucket, flow hash, queue ops, builders, queue service,
cross-binding) extend.

Per the phased plan in docs/pr/956-tx-decomposition/plan.md (v6,
PLAN-READY at Codex round-6 + Gemini AGREE-TO-MERGE-PLAN):

Moved from tx.rs to cos/ecn.rs (~210 LOC):
- 5 ECN codepoint masks (ECN_MASK / NOT_ECT / ECT_0 / ECT_1 / CE)
  declared pub(in crate::afxdp) since admission tests in
  tx::tests still reference them.
- EthernetL3 enum + ethernet_l3 parser (single-VLAN aware,
  rejects QinQ + non-IP payloads). pub(in crate::afxdp) for the
  same reason.
- mark_ecn_ce_ipv4 (RFC 1624 incremental checksum update) +
  mark_ecn_ce_ipv6 (no header checksum).
- maybe_mark_ecn_ce + maybe_mark_ecn_ce_prepared dispatchers.

cos/mod.rs re-exports the marker fns and codepoint masks for the
shorter `super::cos::{...}` import path on tx.rs call sites.

afxdp.rs adds `#[path = "afxdp/cos/mod.rs"] mod cos;` matching
the existing `#[path = "afxdp/tx.rs"] mod tx;` style.

Stays in tx.rs for Phase 1:
- COS_ECN_MARK_THRESHOLD_NUM/_DEN + their two const_assert
  invariants. They are admission-policy tuning knobs, not ECN
  codepoint definitions; moving them into cos::ecn would create
  the wrong dependency direction. They move with admission to
  cos/admission.rs in Phase 2.
- apply_cos_admission_ecn_policy (admission policy entry point).
- All 15 ECN/ethernet_l3 unit tests + the shared test fixtures
  (build_ipv4_test_packet, build_ipv6_test_packet, etc.) — the
  tests reach the moved items via the new use statements. Test
  reorganization deferred to Phase 2 alongside admission move
  (avoids a same-PR fixture relocation while admission tests
  still depend on them).

Net diff: tx.rs -210 LOC, cos/ created. 865 tests pass (no count
change since tests didn't move).
